requirements.txt
elasticsearch>=5.2.0
pydash>=4.1.0
redis>=2.10.5
mysql-connector-python==1.2.3
tornado>=4.4.2
